app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
requests==2.31.0
google-generativeai==0.3.2
mlflow==2.10.0
python-dotenv==1.0.1
packaging==23.2
